discord-bot/requirements.txt
discord.py>=2.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pymongo>=4.5.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.4.0
python-multipart>=0.0.6